@@ -30,22 +62,280 @@
       "prettier --ignore-path .eslintignore --single-quote --write"
     ]
   },
+  "browserslist": [],
+  "prettier": {
+    "singleQuote": true,
+    "overrides": [
+      {
+        "files": [
+          ".prettierrc",
+          ".eslintrc"
+        ],
+        "options": {
+          "parser": "json"
+        }
+      }
+    ]
+  },
+  "jest": {
+    "moduleDirectories": [
+      "node_modules",
+      "release/app/node_modules"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "jsx",
+      "ts",
+      "tsx",
+      "json"
+    ],
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/.erb/mocks/fileMock.js",
+      "\\.(css|less|sass|scss)$": "identity-obj-proxy"
+    },
+    "setupFiles": [
+      "./.erb/scripts/check-build-exists.ts"
+    ],
+    "testEnvironment": "jsdom",
+    "testEnvironmentOptions": {
+      "url": "http://localhost/"
+    },
+    "testPathIgnorePatterns": [
+      "release/app/dist"
+    ],
+    "transform": {
+      "\\.(ts|tsx|js|jsx)$": "ts-jest"
+    }
+  },
+  "dependencies": {
+    "@apollo/client": "^3.4.3",
+    "@date-io/date-fns": "1.3.13",
+    "@fortawesome/fontawesome-free": "^5.15.0",
+    "@iconify/react": "^1.1.4",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/data-grid": "^4.0.0-alpha.35",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@material-ui/pickers": "^3.3.10",
+    "@react-spring/web": "^9.2.4",
+    "@reduxjs/toolkit": "^1.6.1",
+    "@wangeditor/editor": "^5.0.1",
+    "@wangeditor/editor-for-react": "^1.0.2",
+    "axios": "^0.21.1",
+    "chartist": "^0.11.4",
+    "connected-react-router": "^6.9.2",
+    "cp-react-tree-table": "^1.0.0-beta.12",
+    "custom-electron-titlebar": "^4.1.0",
+    "electron-debug": "^3.2.0",
+    "electron-log": "^4.4.7",
+    "electron-updater": "^5.0.1",
+    "emoji-mart": "^3.0.1",
+    "fuse.js": "^6.4.6",
+    "graphql": "^15.5.1",
+    "history": "^4.10.1",
+    "jose": "^4.8.1",
+    "jsoneditor": "^9.5.1",
+    "lodash": "^4.17.15",
+    "lodash.isequal": "^4.5.0",
+    "material-ui-nested-menu-item": "^1.0.2",
+    "normalize.css": "^8.0.1",
+    "perfect-scrollbar": "^1.5.2",
+    "platform": "^1.3.5",
+    "prop-types": "^15.7.2",
+    "pure-render-decorator": "^1.2.1",
+    "rc-upload": "^4.2.0",
+    "re-resizable": "^6.9.0",
+    "react": "^18.1.0",
+    "react-chartist": "^0.14.2",
+    "react-color": "^2.19.3",
+    "react-copy-to-clipboard": "^5.0.2",
+    "react-cropper": "^2.1.8",
+    "react-dom": "^18.1.0",
+    "react-draggable": "^4.4.3",
+    "react-dropdown-tree-select": "^2.5.1",
+    "react-google-maps": "^9.4.5",
+    "react-hook-form": "^7.27.1",
+    "react-hotkeys": "^2.0.0",
+    "react-loading": "^2.0.3",
+    "react-native-web": "^0.17.1",
+    "react-number-format": "^4.6.4",
+    "react-redux": "^8.0.1",
+    "react-router-dom": "^6.3.0",
+    "react-shadow": "^19.0.2",
+    "react-spring": "^9.2.4",
+    "react-swipeable-views": "^0.14.0",
+    "react-viewer": "^3.2.2",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.1",
+    "regenerator-runtime": "^0.13.9",
+    "rxjs": "^7.3.0",
+    "socket.io-client": "^2.3.0",
+    "source-map-support": "^0.5.19",
+    "subscriptions-transport-ws": "^0.9.19",
+    "ts-toolbelt": "^9.6.0",
+    "use-sound": "^4.0.1",
+    "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+    "@amilajack/testcafe-browser-provider-electron": "^0.0.15-alpha.1",
+    "@babel/core": "^7.12.9",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-proposal-decorators": "^7.12.1",
+    "@babel/plugin-proposal-do-expressions": "^7.12.1",
+    "@babel/plugin-proposal-export-default-from": "^7.12.1",
+    "@babel/plugin-proposal-export-namespace-from": "^7.12.1",
+    "@babel/plugin-proposal-function-bind": "^7.12.1",
+    "@babel/plugin-proposal-function-sent": "^7.12.1",
+    "@babel/plugin-proposal-json-strings": "^7.12.1",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.12.1",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.12.1",
+    "@babel/plugin-proposal-optional-chaining": "^7.12.7",
+    "@babel/plugin-proposal-pipeline-operator": "^7.12.1",
+    "@babel/plugin-proposal-throw-expressions": "^7.12.1",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-syntax-import-meta": "^7.10.4",
+    "@babel/plugin-transform-react-constant-elements": "^7.12.1",
+    "@babel/plugin-transform-react-inline-elements": "^7.12.1",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/preset-env": "^7.12.7",
+    "@babel/preset-react": "^7.12.7",
+    "@babel/preset-typescript": "^7.12.7",
+    "@babel/register": "^7.12.1",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.6",
+    "@teamsupercell/typings-for-css-modules-loader": "^2.5.1",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@types/classnames": "^2.2.11",
+    "@types/electron-json-storage": "^4.0.0",
+    "@types/emoji-mart": "^3.0.4",
+    "@types/enzyme": "^3.10.5",
+    "@types/enzyme-adapter-react-16": "^1.0.6",
+    "@types/history": "4.7.6",
+    "@types/jest": "^27.5.1",
+    "@types/jsoneditor": "^8.6.0",
+    "@types/jsonwebtoken": "^8.5.0",
+    "@types/jwk-to-pem": "^2.0.0",
+    "@types/node": "17.0.33",
+    "@types/pure-render-decorator": "^0.2.28",
+    "@types/react": "^18.0.9",
+    "@types/react-color": "^3.0.5",
+    "@types/react-copy-to-clipboard": "^4.3.0",
+    "@types/react-dom": "^18.0.4",
+    "@types/react-native-web": "npm:@types/react-native",
+    "@types/react-redux": "^7.1.9",
+    "@types/react-router": "^5.1.8",
+    "@types/react-router-dom": "^5.1.6",
+    "@types/react-swipeable-views": "^0.13.1",
+    "@types/react-test-renderer": "^18.0.0",
+    "@types/redux-logger": "^3.0.8",
+    "@types/socket.io-client": "^1.4.32",
+    "@types/terser-webpack-plugin": "^5.0.4",
+    "@types/uuid": "^8.3.0",
+    "@types/webpack": "^4.41.21",
+    "@types/webpack-bundle-analyzer": "^4.4.1",
+    "@types/webpack-env": "^1.15.2",
+    "@typescript-eslint/eslint-plugin": "^5.23.0",
+    "@typescript-eslint/parser": "^5.23.0",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^26.1.0",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-dev-expression": "^0.2.2",
+    "babel-plugin-react-native-web": "^0.17.1",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "browserslist-config-erb": "^0.0.3",
+    "chalk": "^4.1.2",
+    "concurrently": "^7.1.0",
+    "core-js": "^3.22.5",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.7.1",
+    "css-minimizer-webpack-plugin": "^3.4.1",
+    "detect-port": "^1.3.0",
+    "electron": "^18.2.3",
+    "electron-builder": "^23.0.3",
+    "electron-devtools-installer": "^3.2.0",
+    "electron-notarize": "^1.2.1",
+    "electron-rebuild": "^3.2.7",
+    "electronmon": "^2.0.2",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.3",
+    "enzyme-to-json": "^3.5.0",
+    "eslint": "^8.15.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^12.0.0",
+    "eslint-config-erb": "^4.0.3",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-import-resolver-webpack": "^0.13.2",
+    "eslint-plugin-compat": "^4.0.2",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.2.2",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.5.0",
+    "eslint-plugin-testcafe": "^0.2.1",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": "^8.0.1",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^28.1.0",
+    "jest-environment-jsdom": "^28.1.0",
+    "less": "^3.11.1",
+    "less-loader": "^5.0.0",
+    "less-plugin-autoprefix": "^2.0.0",
+    "lint-staged": "^12.4.1",
+    "mini-css-extract-plugin": "^2.6.0",
+    "node-sass": "^5.0.0",
+    "opencollective-postinstall": "^2.0.3",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "prettier": "^2.6.2",
+    "react-refresh": "^0.13.0",
+    "react-test-renderer": "^18.1.0",
+    "redux-logger": "^3.0.6",
+    "rimraf": "^3.0.2",
+    "sass": "^1.51.0",
+    "sass-loader": "^12.6.0",
+    "style-loader": "^3.3.1",
+    "stylelint": "^13.6.1",
+    "stylelint-config-prettier": "^8.0.2",
+    "stylelint-config-standard": "^20.0.0",
+    "terser-webpack-plugin": "^5.3.1",
+    "testcafe": "^1.8.8",
+    "testcafe-browser-provider-electron": "^0.0.15",
+    "testcafe-react-selectors": "^4.0.0",
+    "ts-jest": "^28.0.2",
+    "ts-loader": "^9.3.0",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.72.1",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0",
+    "webpack-merge": "^5.8.0",
+    "webpack-mild-compile": "^3.3.1",
+    "yarn-deduplicate": "^3.1.0"
+  },
   "build": {
     "productName": "小白客服",
     "appId": "com.xiaobai.ContactCenterClient",
+    "asar": true,
+    "asarUnpack": "**\\*.{node,dll}",
     "files": [
-      "dist/",
-      "node_modules/",
-      "index.html",
-      "main.prod.js",
-      "main.prod.js.map",
+      "dist",
+      "node_modules",
       "package.json"
     ],
-    "afterSign": ".erb/scripts/Notarize.js",
+    "afterSign": ".erb/scripts/notarize.js",
     "mac": {
-      "target": [
-        "dmg"
-      ],
+      "target": {
+        "target": "default",
+        "arch": [
+          "arm64",
+          "x64"
+        ]
+      },
       "type": "distribution",
       "hardenedRuntime": true,
       "entitlements": "assets/entitlements.mac.plist",
@@ -91,290 +381,18 @@
       "repo": "contact-center-client"
     }
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nedphae/contact-center-client.git"
-  },
-  "author": {
-    "name": "高万里",
-    "email": "nedphae@gmail.com",
-    "url": "https://github.com/nedphae/contact-center"
-  },
-  "contributors": [
-    {
-      "name": "高万里",
-      "email": "nedphae@gmail.com",
-      "url": "https://github.com/nedphae"
-    }
-  ],
-  "license": "Apache 2.0",
-  "bugs": {
-    "url": "https://github.com/nedphae/contact-center-client/issues"
-  },
-  "keywords": [
-    "electron",
-    "boilerplate",
-    "react",
-    "typescript",
-    "ts",
-    "sass",
-    "webpack",
-    "hot",
-    "reload"
-  ],
-  "homepage": "https://github.com/nedphae/contact-center-client#readme",
-  "jest": {
-    "testURL": "http://localhost/",
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/.erb/mocks/fileMock.js",
-      "\\.(css|less|sass|scss)$": "identity-obj-proxy"
-    },
-    "moduleFileExtensions": [
-      "js",
-      "jsx",
-      "ts",
-      "tsx",
-      "json"
-    ],
-    "moduleDirectories": [
-      "node_modules",
-      "app/node_modules"
-    ],
-    "setupFiles": [
-      "./.erb/scripts/CheckBuildsExist.js"
-    ]
-  },
-  "devDependencies": {
-    "@amilajack/testcafe-browser-provider-electron": "^0.0.15-alpha.1",
-    "@babel/core": "^7.12.9",
-    "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/plugin-proposal-decorators": "^7.12.1",
-    "@babel/plugin-proposal-do-expressions": "^7.12.1",
-    "@babel/plugin-proposal-export-default-from": "^7.12.1",
-    "@babel/plugin-proposal-export-namespace-from": "^7.12.1",
-    "@babel/plugin-proposal-function-bind": "^7.12.1",
-    "@babel/plugin-proposal-function-sent": "^7.12.1",
-    "@babel/plugin-proposal-json-strings": "^7.12.1",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.12.1",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.12.1",
-    "@babel/plugin-proposal-optional-chaining": "^7.12.7",
-    "@babel/plugin-proposal-pipeline-operator": "^7.12.1",
-    "@babel/plugin-proposal-throw-expressions": "^7.12.1",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@babel/plugin-transform-react-constant-elements": "^7.12.1",
-    "@babel/plugin-transform-react-inline-elements": "^7.12.1",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.12.7",
-    "@babel/preset-react": "^7.12.7",
-    "@babel/preset-typescript": "^7.12.7",
-    "@babel/register": "^7.12.1",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
-    "@teamsupercell/typings-for-css-modules-loader": "^2.4.0",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@types/classnames": "^2.2.11",
-    "@types/electron-json-storage": "^4.0.0",
-    "@types/emoji-mart": "^3.0.4",
-    "@types/enzyme": "^3.10.5",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/history": "4.7.6",
-    "@types/jest": "^26.0.15",
-    "@types/jsoneditor": "^8.6.0",
-    "@types/jsonwebtoken": "^8.5.0",
-    "@types/jwk-to-pem": "^2.0.0",
-    "@types/node": "14.14.10",
-    "@types/pure-render-decorator": "^0.2.28",
-    "@types/react": "^16.9.44",
-    "@types/react-color": "^3.0.5",
-    "@types/react-copy-to-clipboard": "^4.3.0",
-    "@types/react-dom": "^16.9.9",
-    "@types/react-native-web": "npm:@types/react-native",
-    "@types/react-redux": "^7.1.9",
-    "@types/react-router": "^5.1.8",
-    "@types/react-router-dom": "^5.1.6",
-    "@types/react-swipeable-views": "^0.13.1",
-    "@types/react-test-renderer": "^16.9.3",
-    "@types/redux-logger": "^3.0.8",
-    "@types/socket.io-client": "^1.4.32",
-    "@types/uuid": "^8.3.0",
-    "@types/webpack": "^4.41.21",
-    "@types/webpack-env": "^1.15.2",
-    "@typescript-eslint/eslint-plugin": "^4.8.1",
-    "@typescript-eslint/parser": "^4.8.1",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.1.0",
-    "babel-loader": "^8.2.2",
-    "babel-plugin-dev-expression": "^0.2.2",
-    "babel-plugin-react-native-web": "^0.17.1",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "browserslist-config-erb": "^0.0.1",
-    "chalk": "^4.1.0",
-    "concurrently": "^5.3.0",
-    "core-js": "^3.6.5",
-    "cross-env": "^7.0.2",
-    "css-loader": "^5.0.1",
-    "css-minimizer-webpack-plugin": "^1.1.5",
-    "detect-port": "^1.3.0",
-    "electron": "^15.1.0",
-    "electron-builder": "^22.11.7",
-    "electron-devtools-installer": "^3.2.0",
-    "electron-notarize": "^1.1.1",
-    "electron-rebuild": "^3.2.3",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.3",
-    "enzyme-to-json": "^3.5.0",
-    "eslint": "^7.5.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-airbnb-typescript": "^12.0.0",
-    "eslint-config-erb": "^2.0.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-import-resolver-webpack": "^0.13.0",
-    "eslint-plugin-compat": "^3.8.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jest": "^24.1.3",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.20.6",
-    "eslint-plugin-react-hooks": "^4.0.8",
-    "eslint-plugin-testcafe": "^0.2.1",
-    "file-loader": "^6.0.0",
-    "husky": "^4.2.5",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.1.0",
-    "less": "^3.11.1",
-    "less-loader": "^5.0.0",
-    "less-plugin-autoprefix": "^2.0.0",
-    "lint-staged": "^10.2.11",
-    "mini-css-extract-plugin": "^1.3.1",
-    "node-sass": "^5.0.0",
-    "opencollective-postinstall": "^2.0.3",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "prettier": "^2.0.5",
-    "react-refresh": "^0.9.0",
-    "react-test-renderer": "^17.0.1",
-    "redux-logger": "^3.0.6",
-    "rimraf": "^3.0.0",
-    "sass-loader": "^10.1.0",
-    "style-loader": "^2.0.0",
-    "stylelint": "^13.6.1",
-    "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-standard": "^20.0.0",
-    "terser-webpack-plugin": "^5.0.3",
-    "testcafe": "^1.8.8",
-    "testcafe-browser-provider-electron": "^0.0.15",
-    "testcafe-react-selectors": "^4.0.0",
-    "typescript": "^4.1.0",
-    "url-loader": "^4.1.0",
-    "webpack": "^5.5.1",
-    "webpack-bundle-analyzer": "^4.1.0",
-    "webpack-cli": "^4.2.0",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^5.4.0",
-    "webpack-mild-compile": "^3.3.1",
-    "yarn-deduplicate": "^3.1.0"
-  },
-  "dependencies": {
-    "@apollo/client": "^3.4.3",
-    "@date-io/date-fns": "1.3.13",
-    "@fortawesome/fontawesome-free": "^5.15.0",
-    "@iconify/react": "^1.1.4",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/data-grid": "^4.0.0-alpha.35",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@material-ui/pickers": "^3.3.10",
-    "@react-spring/web": "^9.2.4",
-    "@reduxjs/toolkit": "^1.6.1",
-    "@wangeditor/editor": "^5.0.1",
-    "@wangeditor/editor-for-react": "^1.0.2",
-    "axios": "^0.21.1",
-    "chartist": "^0.11.4",
-    "connected-react-router": "^6.6.1",
-    "cp-react-tree-table": "^1.0.0-beta.12",
-    "electron-debug": "^3.2.0",
-    "electron-json-storage": "^4.3.0",
-    "electron-log": "^4.4.1",
-    "electron-updater": "^4.3.4",
-    "emoji-mart": "^3.0.1",
-    "fuse.js": "^6.4.6",
-    "graphql": "^15.5.1",
-    "history": "^4.10.1",
-    "jsoneditor": "^9.5.1",
-    "jsonwebtoken": "^8.5.1",
-    "jwks-rsa": "^2.0.4",
-    "lodash": "^4.17.15",
-    "lodash.isequal": "^4.5.0",
-    "material-ui-nested-menu-item": "^1.0.2",
-    "normalize.css": "^8.0.1",
-    "perfect-scrollbar": "^1.5.2",
-    "platform": "^1.3.5",
-    "prop-types": "^15.7.2",
-    "pure-render-decorator": "^1.2.1",
-    "rc-upload": "^4.2.0",
-    "re-resizable": "^6.9.0",
-    "react": "^17.0.2",
-    "react-chartist": "^0.14.2",
-    "react-color": "^2.19.3",
-    "react-copy-to-clipboard": "^5.0.2",
-    "react-cropper": "^2.1.8",
-    "react-dom": "^17.0.2",
-    "react-draggable": "^4.4.3",
-    "react-dropdown-tree-select": "^2.5.1",
-    "react-google-maps": "^9.4.5",
-    "react-hook-form": "^7.27.1",
-    "react-hotkeys": "^2.0.0",
-    "react-loading": "^2.0.3",
-    "react-native-web": "^0.17.1",
-    "react-number-format": "^4.6.4",
-    "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
-    "react-shadow": "^19.0.2",
-    "react-spring": "^9.2.4",
-    "react-swipeable-views": "^0.14.0",
-    "react-viewer": "^3.2.2",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "regenerator-runtime": "^0.13.9",
-    "rxjs": "^7.3.0",
-    "socket.io-client": "^2.3.0",
-    "source-map-support": "^0.5.19",
-    "subscriptions-transport-ws": "^0.9.19",
-    "ts-toolbelt": "^9.6.0",
-    "use-sound": "^4.0.1",
-    "uuid": "^8.3.2"
+  "collective": {
+    "url": "https://opencollective.com/electron-react-boilerplate-594"
   },
   "devEngines": {
-    "node": ">=10.x",
-    "npm": ">=6.x",
-    "yarn": ">=1.21.3"
+    "node": ">=14.x",
+    "npm": ">=7.x"
   },
-  "collective": {
-    "url": "https://github.com/nedphae"
-  },
-  "browserslist": [],
-  "prettier": {
-    "overrides": [
-      {
-        "files": [
-          ".prettierrc",
-          ".babelrc",
-          ".eslintrc"
-        ],
-        "options": {
-          "parser": "json"
-        }
-      }
+  "electronmon": {
+    "patterns": [
+      "!**/**",
+      "src/main/*"
     ],
-    "singleQuote": true
-  },
-  "renovate": {
-    "extends": [
-      "bliss"
-    ],
-    "baseBranches": [
-      "next"
-    ]
+    "logLevel": "quiet"
   }
 }
